requirements.txt
fastapi==0.93.0
pydantic==1.10.6
uvicorn==0.21.0
Requests==2.31.0
numpy==1.24.3
imutils==0.5.4
dlib==19.24.2
matplotlib==3.7.1
opencv-python==4.6.0.66
scipy==1.10.1
python-dotenv==1.0.0
